+++ Code/LYRUIAvatarImageView.h
@@ -10,9 +10,8 @@
 
 @interface LYRUIAvatarImageView : UIImageView
 
-@property (nonatomic) UIFont *initialFont UI_APPEARANCE_SELECTOR;
-@property (nonatomic) UIColor *initialColor UI_APPEARANCE_SELECTOR;
-@property (nonatomic) UIColor *initialViewBackgroundColor UI_APPEARANCE_SELECTOR;
+@property (nonatomic) UIFont *initialsFont UI_APPEARANCE_SELECTOR;
+@property (nonatomic) UIColor *initialsColor UI_APPEARANCE_SELECTOR;
 
 - (void)setInitialsForName:(NSString *)name;