+++ Code/LYRUIIncomingMessageCollectionViewCell.m
@@ -10,16 +10,8 @@
 
 @interface LYRUIIncomingMessageCollectionViewCell ()
 
-@property (nonatomic) UIFont *font;
-@property (nonatomic) UIColor *color;
-@property (nonatomic) CGFloat imageViewDiameter;
-@property (nonatomic) CGFloat imageViewLeft;
-@property (nonatomic) BOOL avatarConstraintsAreSet;
-
-@property (nonatomic) NSLayoutConstraint *avatarImageWidthConstraint;
-@property (nonatomic) NSLayoutConstraint *avatarImageHeightConstraint;
-@property (nonatomic) NSLayoutConstraint *avatarImageBottomConstraint;
-@property (nonatomic) NSLayoutConstraint *avatarImageLeftConstraint;
+@property (nonatomic) NSLayoutConstraint *bubbleWithAvatarLeftConstraint;
+@property (nonatomic) NSLayoutConstraint *bubbleWithoutAvatarLeftConstraint;
 
 @end
 
@@ -31,100 +23,88 @@ static CGFloat const LYRAvatarImageDiameter = 30.0f;
 {
     self = [super initWithFrame:frame];
     if (self) {
+        self.avatarImageView.layer.cornerRadius = (LYRAvatarImageDiameter / 2);
+        self.avatarImageView.clipsToBounds = YES;
         
-        self.avatarImage.layer.cornerRadius = (LYRAvatarImageDiameter / 2);
-        self.avatarConstraintsAreSet = NO;
-        
-        [self updateAvatarImageConstraints];
-        
+        [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView
+                                                                     attribute:NSLayoutAttributeWidth
+                                                                     relatedBy:NSLayoutRelationEqual
+                                                                        toItem:nil
+                                                                     attribute:NSLayoutAttributeNotAnAttribute
+                                                                    multiplier:1.0
+                                                                      constant:LYRAvatarImageDiameter]];
+
+        [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView
+                                                                     attribute:NSLayoutAttributeHeight
+                                                                     relatedBy:NSLayoutRelationEqual
+                                                                        toItem:self.avatarImageView
+                                                                     attribute:NSLayoutAttributeWidth
+                                                                    multiplier:1.0
+                                                                      constant:0]];
+
+        [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView
+                                                                     attribute:NSLayoutAttributeLeft
+                                                                     relatedBy:NSLayoutRelationEqual
+                                                                        toItem:self.contentView
+                                                                     attribute:NSLayoutAttributeLeft
+                                                                    multiplier:1.0
+                                                                      constant:10]];
+
+        [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.avatarImageView
+                                                                     attribute:NSLayoutAttributeBottom
+                                                                     relatedBy:NSLayoutRelationEqual
+                                                                        toItem:self.contentView
+                                                                     attribute:NSLayoutAttributeBottom
+                                                                    multiplier:1.0
+                                                                      constant:0]];
+
+        self.bubbleWithAvatarLeftConstraint = [NSLayoutConstraint constraintWithItem:self.bubbleView
+                                                                           attribute:NSLayoutAttributeLeft
+                                                                           relatedBy:NSLayoutRelationEqual
+                                                                              toItem:self.avatarImageView
+                                                                           attribute:NSLayoutAttributeRight
+                                                                          multiplier:1.0
+                                                                            constant:10];
+        [self.contentView addConstraint:self.bubbleWithAvatarLeftConstraint];
+
+        self.bubbleWithoutAvatarLeftConstraint = [NSLayoutConstraint constraintWithItem:self.bubbleView
+                                                                              attribute:NSLayoutAttributeLeft
+                                                                              relatedBy:NSLayoutRelationEqual
+                                                                                 toItem:self.avatarImageView
+                                                                              attribute:NSLayoutAttributeLeft
+                                                                             multiplier:1.0
+                                                                               constant:0];
     }
     return self;
 }
 
 - (void)shouldDisplayAvatarImage:(BOOL)shouldDisplayAvatarImage
 {
-    if (!shouldDisplayAvatarImage) {
-        self.avatarImage.alpha = 1.0;
-        self.avatarImageLeftConstraint.constant = 0;
-        self.avatarImageWidthConstraint.constant = 0;
-        self.avatarImageHeightConstraint.constant = 0;
-
+    NSArray *constraints = [self.contentView constraints];
+    if (shouldDisplayAvatarImage) {
+        if ([constraints containsObject:self.bubbleWithAvatarLeftConstraint]) return;
+        [self.contentView removeConstraint:self.bubbleWithoutAvatarLeftConstraint];
+        [self.contentView addConstraint:self.bubbleWithAvatarLeftConstraint];
     } else {
-        self.avatarImage.alpha = 0.0;
-        self.avatarImageLeftConstraint.constant = 10;
-        self.avatarImageWidthConstraint.constant = LYRAvatarImageDiameter;
-        self.avatarImageHeightConstraint.constant = LYRAvatarImageDiameter;
+        if ([constraints containsObject:self.bubbleWithoutAvatarLeftConstraint]) return;
+        [self.contentView removeConstraint:self.bubbleWithAvatarLeftConstraint];
+        [self.contentView addConstraint:self.bubbleWithoutAvatarLeftConstraint];
     }
-    [self.contentView layoutIfNeeded];
+    [self setNeedsUpdateConstraints];
 }
 
 - (void)updateWithParticipant:(id<LYRUIParticipant>)participant
 {
     if (participant) {
-        self.avatarImage.alpha = 1.0f;
-        self.avatarImage.layer.cornerRadius = LYRAvatarImageDiameter / 2;
-        self.avatarImage.clipsToBounds = YES;
+        self.avatarImageView.hidden = NO;
         if (participant.avatarImage) {
-            [self.avatarImage setImage:participant.avatarImage];
+            [self.avatarImageView setImage:participant.avatarImage];
         } else {
-            [self.avatarImage setInitialsForName:participant.fullName];
+            [self.avatarImageView setInitialsForName:participant.fullName];
         }
     } else {
-        self.avatarImage.alpha = 0.0f;
+        self.avatarImageView.hidden = YES;
     }
-}
-
-- (void)updateAvatarImageConstraints
-{
-    //***************Avatar Image Constraints***************//
-    self.avatarImageWidthConstraint = [NSLayoutConstraint constraintWithItem:self.avatarImage
-                                                                   attribute:NSLayoutAttributeWidth
-                                                                   relatedBy:NSLayoutRelationEqual
-                                                                      toItem:nil
-                                                                   attribute:NSLayoutAttributeNotAnAttribute
-                                                                  multiplier:1.0
-                                                                    constant:self.imageViewDiameter];
-    
-    self.avatarImageHeightConstraint = [NSLayoutConstraint constraintWithItem:self.avatarImage
-                                                                    attribute:NSLayoutAttributeHeight
-                                                                    relatedBy:NSLayoutRelationEqual
-                                                                       toItem:nil
-                                                                    attribute:NSLayoutAttributeNotAnAttribute
-                                                                   multiplier:1.0
-                                                                     constant:self.imageViewDiameter];
-    
-    self.avatarImageBottomConstraint = [NSLayoutConstraint constraintWithItem:self.avatarImage
-                                                                    attribute:NSLayoutAttributeBottom
-                                                                    relatedBy:NSLayoutRelationEqual
-                                                                       toItem:self.contentView
-                                                                    attribute:NSLayoutAttributeBottom
-                                                                   multiplier:1.0
-                                                                     constant:0];
-    
-    self.avatarImageLeftConstraint = [NSLayoutConstraint constraintWithItem:self.avatarImage
-                                                                  attribute:NSLayoutAttributeLeft
-                                                                  relatedBy:NSLayoutRelationEqual
-                                                                     toItem:self.contentView
-                                                                  attribute:NSLayoutAttributeLeft
-                                                                 multiplier:1.0
-                                                                   constant:self.imageViewLeft];
-    // Add avatar constraints
-    [self.contentView addConstraint:self.avatarImageWidthConstraint];
-    [self.contentView addConstraint:self.avatarImageHeightConstraint];
-    [self.contentView addConstraint:self.avatarImageBottomConstraint];
-    [self.contentView addConstraint:self.avatarImageLeftConstraint];
-}
-
-- (void)updateConstraints
-{
-    [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.bubbleView
-                                                                 attribute:NSLayoutAttributeLeft
-                                                                 relatedBy:NSLayoutRelationEqual
-                                                                    toItem:self.avatarImage
-                                                                 attribute:NSLayoutAttributeRight
-                                                                multiplier:1.0
-                                                                  constant:10]];
-    [super updateConstraints];
 }
 
 @end