+++ Code/LYRUIMessageBubbleView.h
@@ -11,34 +11,34 @@
 
 /**
  @abstract The `LYRUIMessageBubbleView` class provides a lightweight, customizable view that 
- handles displaying the actual message content within a collectionViewCell. 
- @discussion The view provides support for multiple different content types including text,
+ handles displaying the actual message content within a collection view cell.
+ @discussion The view provides support for multiple content types including text,
  images, and location data.
  */
 @interface LYRUIMessageBubbleView : UIView <UIAppearanceContainer>
 
 /**
- @abstract Tells the bubble view to display a given string
+ @abstract Tells the bubble view to display a given string.
  */
 - (void)updateWithText:(NSString *)text;
 
 /**
- @abstract Tells the bubble view to display a given image
+ @abstract Tells the bubble view to display a given image.
  */
 - (void)updateWithImage:(UIImage *)image;
 
 /**
- @abstract Tells the bubble view to display a map image for a given location
+ @abstract Tells the bubble view to display a map image for a given location.
  */
 - (void)updateWithLocation:(CLLocationCoordinate2D)location;
 
 /**
- @abstract The textView object that handles displaying text
+ @abstract The view that handles displaying text.
  */
 @property (nonatomic) UILabel *bubbleViewLabel;
 
 /**
- @abstract imageView object that handles displaying images
+ @abstract The view that handles displaying an image.
  */
 @property (nonatomic) UIImageView *bubbleImageView;