+++ Code/LYRUIMessageCollectionViewCell.m
@@ -13,16 +13,8 @@
 
 @interface LYRUIMessageCollectionViewCell ()
 
-@property (nonatomic) UILabel *dateLabel;
-@property (nonatomic) CGFloat bubbleViewWidth;
 @property (nonatomic) BOOL messageSentState;
-
-@property (nonatomic) NSLayoutConstraint *bubbleViewHeightConstraint;
-@property (nonatomic) NSLayoutConstraint *bubbleViewTopConstraint;
 @property (nonatomic) NSLayoutConstraint *bubbleViewWidthConstraint;
-
-@property (nonatomic) NSLayoutConstraint *dateLabelLeftConstraint;
-@property (nonatomic) NSLayoutConstraint *dateLabelCenterYConstraint;
 
 @end
 
@@ -32,43 +24,56 @@
 {
     self = [super initWithFrame:frame];
     if (self) {
-        
         self.bubbleView = [[LYRUIMessageBubbleView alloc] init];
         self.bubbleView.translatesAutoresizingMaskIntoConstraints = NO;
         [self.contentView addSubview:self.bubbleView];
         
-        self.avatarImage = [[LYRUIAvatarImageView alloc] init];
+        self.avatarImageView = [[LYRUIAvatarImageView alloc] init];
+        self.avatarImageView.backgroundColor = LSLighGrayColor();
+        self.avatarImageView.translatesAutoresizingMaskIntoConstraints = NO;
+        [self.contentView addSubview:self.avatarImageView];
 
-        [self.avatarImage setInitialViewBackgroundColor:LSLighGrayColor()];
-        self.avatarImage.translatesAutoresizingMaskIntoConstraints = NO;
-        [self.contentView addSubview:self.avatarImage];
-       
-        self.dateLabel  = [[UILabel alloc] init];
-        self.dateLabel.font = [UIFont systemFontOfSize:10];
-        self.dateLabel.textColor = [UIColor lightGrayColor];
-        self.dateLabel.text = @"10:36AM";
-        self.dateLabel.translatesAutoresizingMaskIntoConstraints = NO;
-        [self.dateLabel sizeToFit];
-        [self.contentView addSubview:self.dateLabel];
+        [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.bubbleView
+                                                                     attribute:NSLayoutAttributeHeight
+                                                                     relatedBy:NSLayoutRelationEqual
+                                                                        toItem:self.contentView
+                                                                     attribute:NSLayoutAttributeHeight
+                                                                    multiplier:1.0
+                                                                      constant:0]];
 
+        [self.contentView addConstraint:[NSLayoutConstraint constraintWithItem:self.bubbleView
+                                                                     attribute:NSLayoutAttributeTop
+                                                                     relatedBy:NSLayoutRelationEqual
+                                                                        toItem:self.contentView
+                                                                     attribute:NSLayoutAttributeTop
+                                                                    multiplier:1.0
+                                                                      constant:0]];
+
+        self.bubbleViewWidthConstraint = [NSLayoutConstraint constraintWithItem:self.bubbleView
+                                                                      attribute:NSLayoutAttributeWidth
+                                                                      relatedBy:NSLayoutRelationEqual
+                                                                         toItem:nil
+                                                                      attribute:NSLayoutAttributeNotAnAttribute
+                                                                     multiplier:1.0
+                                                                       constant:0];
+        [self.contentView addConstraint:self.bubbleViewWidthConstraint];
     }
-    [self updateMessageCellConstraints];
     return self;
 }
 
 - (void)updateWithParticipant:(id<LYRUIParticipant>)participant
 {
-    NSLog(@"Implemented by subclass");
+
 }
 
 - (void)shouldDisplayAvatarImage:(BOOL)shouldDisplayAvatarImage
 {
-    NSLog(@"Implemented by subclass");
+
 }
 
 - (void)isGroupConversation:(BOOL)isGroupConversation
 {
-    NSLog(@"Implemented by subclass");
+
 }
 
 - (void)presentMessagePart:(LYRMessagePart *)messagePart
@@ -77,7 +82,7 @@
         NSString *text = [[NSString alloc] initWithData:messagePart.data encoding:NSUTF8StringEncoding];
         [self.bubbleView updateWithText:text];
         self.accessibilityLabel = [NSString stringWithFormat:@"Message: %@", text];
-    } else if ([messagePart.MIMEType isEqualToString:LYRUIMIMETypeImageJPEG] || [messagePart.MIMEType isEqualToString:LYRUIMIMETypeImageJPEG]) {
+    } else if ([messagePart.MIMEType isEqualToString:LYRUIMIMETypeImageJPEG] || [messagePart.MIMEType isEqualToString:LYRUIMIMETypeImagePNG]) {
         UIImage *image = [UIImage imageWithData:messagePart.data];
         [self.bubbleView updateWithImage:image];
         self.accessibilityLabel = [NSString stringWithFormat:@"Message: Photo"];
@@ -85,8 +90,8 @@
         NSDictionary *dictionary = [NSJSONSerialization JSONObjectWithData:messagePart.data
                                                                    options:NSJSONReadingAllowFragments
                                                                      error:nil];
-        double lat = [[dictionary valueForKey:@"lat"] doubleValue];
-        double lon = [[dictionary valueForKey:@"lon"] doubleValue];
+        double lat = [dictionary[@"lat"] doubleValue];
+        double lon = [dictionary[@"lon"] doubleValue];
         [self.bubbleView updateWithLocation:CLLocationCoordinate2DMake(lat, lon)];
     }
 }
@@ -98,112 +103,25 @@
 
 - (void)updateWithBubbleViewWidth:(CGFloat)bubbleViewWidth
 {
-    if ([[self.contentView constraints] containsObject:self.bubbleViewWidthConstraint]) {
-        [self.contentView removeConstraint:self.bubbleViewWidthConstraint];
-    }
-    self.bubbleViewWidth = bubbleViewWidth + 26; //Adding 24px for bubble view horizontal padding + 2px for extra coverage
-    self.bubbleViewWidthConstraint = [NSLayoutConstraint constraintWithItem:self.bubbleView
-                                                                  attribute:NSLayoutAttributeWidth
-                                                                  relatedBy:NSLayoutRelationEqual
-                                                                     toItem:nil
-                                                                  attribute:NSLayoutAttributeNotAnAttribute
-                                                                 multiplier:1.0
-                                                                   constant:self.bubbleViewWidth];
-     [self.contentView addConstraint:self.bubbleViewWidthConstraint];
-}
-
-
-- (void)updateMessageCellConstraints
-{
-    //***************Bubble View Constraints***************//
-    self.bubbleViewHeightConstraint = [NSLayoutConstraint constraintWithItem:self.bubbleView
-                                                                    attribute:NSLayoutAttributeHeight
-                                                                    relatedBy:NSLayoutRelationEqual
-                                                                       toItem:self.contentView
-                                                                    attribute:NSLayoutAttributeHeight
-                                                                   multiplier:1.0
-                                                                     constant:0];
-    
-    self.bubbleViewTopConstraint = [NSLayoutConstraint constraintWithItem:self.bubbleView
-                                                                attribute:NSLayoutAttributeTop
-                                                                relatedBy:NSLayoutRelationEqual
-                                                                   toItem:self.contentView
-                                                                attribute:NSLayoutAttributeTop
-                                                               multiplier:1.0
-                                                                 constant:0];
-    //***************Date Label Constraints***************//
-    self.dateLabelLeftConstraint = [NSLayoutConstraint constraintWithItem:self.dateLabel
-                                                                attribute:NSLayoutAttributeLeft
-                                                                relatedBy:NSLayoutRelationEqual
-                                                                   toItem:self.contentView
-                                                                attribute:NSLayoutAttributeRight
-                                                               multiplier:1.0
-                                                                 constant:2];
-    
-    self.dateLabelCenterYConstraint = [NSLayoutConstraint constraintWithItem:self.dateLabel
-                                                                   attribute:NSLayoutAttributeCenterY
-                                                                   relatedBy:NSLayoutRelationEqual
-                                                                      toItem:self.contentView
-                                                                   attribute:NSLayoutAttributeCenterY
-                                                                  multiplier:1.0
-                                                                    constant:0];
-    
-    // Add bubbleView constraints
-    [self.contentView addConstraint:self.bubbleViewHeightConstraint];
-    [self.contentView addConstraint:self.bubbleViewTopConstraint];
-    
-    // Add date label constraints
-    [self.contentView addConstraint:self.dateLabelLeftConstraint];
-    [self.contentView addConstraint:self.dateLabelCenterYConstraint];
+    self.bubbleViewWidthConstraint.constant = bubbleViewWidth + 26; // Adding 24px for bubble view horizontal padding + 2px for extra coverage.
 }
 
 - (void)setMessageTextFont:(UIFont *)messageTextFont
 {
+    _messageTextFont = messageTextFont;
     self.bubbleView.bubbleViewLabel.font = messageTextFont;
 }
 
 - (void)setMessageTextColor:(UIColor *)messageTextColor
 {
+    _messageTextColor = messageTextColor;
     self.bubbleView.bubbleViewLabel.textColor = messageTextColor;
 }
 
 - (void)setBubbleViewColor:(UIColor *)bubbleViewColor
 {
-    if ([self isKindOfClass:[LYRUIOutgoingMessageCollectionViewCell class]] && self.messageSentState) {
-        self.bubbleView.backgroundColor = bubbleViewColor;
-    }
-    if ([self isKindOfClass:[LYRUIIncomingMessageCollectionViewCell class]]) {
-        self.bubbleView.backgroundColor = bubbleViewColor;
-    }
+    _bubbleViewColor = bubbleViewColor;
+    self.bubbleView.backgroundColor = bubbleViewColor;
 }
 
-- (void)setPendingBubbleViewColor:(UIColor *)pendingBubbleViewColor
-{
-    if ([self isKindOfClass:[LYRUIOutgoingMessageCollectionViewCell class]] && !self.messageSentState) {
-        self.bubbleView.backgroundColor = pendingBubbleViewColor;
-    }
-}
-
-- (UIFont *)messageTextFont
-{
-    return self.bubbleView.bubbleViewLabel.font;
-}
-
-- (UIColor *)messageTextColor
-{
-    return self.bubbleView.bubbleViewLabel.textColor;
-}
-
-- (UIColor *)bubbleViewColor
-{
-    return self.bubbleView.backgroundColor;
-}
-
-- (void)layoutSubviews
-{
-    [super layoutSubviews];
-    if ([self isKindOfClass:[LYRUIIncomingMessageCollectionViewCell class]]) {
-        //UIBezierPath
-    }
-}
 @end